Reorganize tests.
  - Reorganize automated tests into multiple files.
  - Rename test case objects to make it clearer than they inherit from
    FacebookTestCase.
  - Remove broken `test_request` test case.
